% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCovariances3DAnemometer.R
\name{calculateCovariances3DAnemometer}
\alias{calculateCovariances3DAnemometer}
\title{calculateCovariances3DAnemometer}
\usage{
calculateCovariances3DAnemometer(windDataframe, timeVariableName = "time",
  UVariableName = "U", VVariableName = "V", WVariableName = "W",
  SonicTemperatureVariableName = "T", timeIntervalType = "ceiling",
  timeIntervalValue = 15, timeIntervalUnits = "minutes")
}
\arguments{
\item{windDataframe}{the data frame for wind data}

\item{timeVariableName}{the name of the time variable}

\item{UVariableName}{the column name of wind in the X direction}

\item{VVariableName}{the column name of wind in the Y direction}

\item{WVariableName}{the column name of wind in the Z direction}

\item{SonicTemperatureVariableName}{the column name of the sonic temperature}

\item{timeIntervalType}{is set to "ceiling" or "floor" depending on where to mark the samples}

\item{timeIntervalValue}{the value which describes the length of the time segment}

\item{timeIntervalUnits}{the units for the length of the time segment}
}
\description{
This function is designed to accept raw data from a 3D anemometer and 
divide the data into time segments.  This data should consist of a 
time, a verticle and two cardinal horizontal wind directions.
}
\details{
For each time segment the
}
\examples{
 
   calculateCovariances3DAnemometer()
}
\keyword{3DAnemometer}
\keyword{time}
