% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMeansOverTime.R
\name{calculateMeansOverTime}
\alias{calculateMeansOverTime}
\title{calculateMeansOverTime}
\usage{
calculateMeansOverTime(dataframe, timeColumnName = "time",
  valueColumnName = NULL, timeIntervalType = "ceiling",
  timeIntervalValue = 15, timeIntervalUnits = "minutes")
}
\arguments{
\item{dataframe}{the data frame for wind data}

\item{timeColumnName}{the name of the time variable}

\item{valueColumnName}{the value of the dataframe to perform statistics}

\item{timeIntervalType}{is set to "ceiling" or "floor" depending on where to mark the samples}

\item{timeIntervalValue}{the value which describes the length of the time segment}

\item{timeIntervalUnits}{the units for the length of the time segment}
}
\description{
This function is designed to accept raw data and 
divide the data into time segments.
}
\details{
For each time segment the outputs are:

mean - mean of each time segment
   stdDev - standard deviation of each time segment
   count - count of each time segment
   max - max of each time segment
   min - min of each time segment
}
\examples{
 
   calculateMeansOverTime()
}
\keyword{means}
\keyword{time}
