% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSonicTemperature.R
\name{calculateSonicTemperature}
\alias{calculateSonicTemperature}
\title{calculateSonicTemperature}
\usage{
calculateSonicTemperature(speedOfSound, GammaConstant = 1.4, Rd = 287.04)
}
\arguments{
\item{speedOfSound}{the speed of sound in m/s}

\item{GammaConstant}{ratio of specific heat of dry air at constant pressure to that at a constant volume}

\item{Rd}{ideal gas constant for dry air (units J K^-1 kg^-1)}
}
\description{
This function calculates the sonic virtual temperature in air based on the measured speed of sound.
}
\details{
Results are in degrees Celcius.
}
\examples{

}
\keyword{emissions}
