% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatTime.R
\name{reformatTime}
\alias{reformatTime}
\title{reformatTime}
\usage{
reformatTime(cdate, ctime = NULL, format = "\%m/\%d/\%Y \%H:\%M:\%S",
  tz = "Etc/GMT+6")
}
\arguments{
\item{cdate}{the date an event occured}

\item{ctime}{the time of an event}

\item{format}{format string to output results}

\item{tz}{time zone to use}
}
\description{
This function reformats time so it is always in a consistant format.
}
\details{
This function is useful for importing date from a text file
}
\examples{

}
\keyword{time}
