% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDataframeColumns.R
\name{removeDataframeColumns}
\alias{removeDataframeColumns}
\title{removeDataframeColumns}
\usage{
removeDataframeColumns(dataframe, ...)
}
\arguments{
\item{dataframe}{the dataframe to removed the columns from}

\item{...}{the column names to be removed as character strings (quoted).}
}
\value{
same dataframe with columns removed
}
\description{
This function is removes unneeded columns from datasets.
}
\examples{

df <- data.frame(column1 = 1, column2 = 2, column3 = 3)
df <- removeDataframeColumns(df, 'column1', 'column3')

}
\keyword{Columns}
\keyword{Dataframes}
