% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameAllDataframeColumns.R
\name{renameAllDataframeColumns}
\alias{renameAllDataframeColumns}
\title{renameAllDataframeColumns}
\usage{
renameAllDataframeColumns(dataframe, newColumnNames, truncateColumns = FALSE,
  truncateStartColumn = 1)
}
\arguments{
\item{dataframe}{the dataframe to edit}

\item{newColumnNames}{the new column names}

\item{truncateColumns}{if TRUE this function removes extra columns not in the column names}

\item{truncateStartColumn}{starting column number truncation starts}
}
\value{
same dataframe except with columns renamed
}
\description{
This function renames all the columns in a dataframe.
}
\details{
This function is useful because wind measuring equipment may output a less than
ideal format for R.  If column names are less than ideal for use in R 
this function can be used to change the names of all of the columns.

The first parameter is the dataframe

The second parameter is the column names to change the header.

If truncate columns is true then extra columns are removed.  This is useful
if a header does not import properly.
}
\examples{

df <- data.frame(Column1 = 1, Column2 = 2)
df <- renameAllDataframeColumns(df, c('ColumnA', 'ColumnB'))



}
\keyword{column}
\keyword{dataframe}
\keyword{rename}
