% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameDataframeColumn.R
\name{renameDataframeColumn}
\alias{renameDataframeColumn}
\title{renameDataframeColumn}
\usage{
renameDataframeColumn(dataframe, oldColumnName, newColumnName)
}
\arguments{
\item{dataframe}{the dataframe to edit}

\item{oldColumnName}{the old column name (Can also be the number of the column)}

\item{newColumnName}{the new name to change the old name to}
}
\value{
same dataframe except with columns renamed
}
\description{
This function renames columns in R.
}
\examples{

df <- data.frame(Column1 = 1, Column2 = 2)
renameDataframeColumn(df, 'Column1', 'ColumnX')
renameDateframeColumn(df, 2, 'ColumnY')

}
\keyword{column}
\keyword{dataframe}
\keyword{rename}
