% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeCeiling.R
\name{timeCeiling}
\alias{timeCeiling}
\title{timeCeiling}
\usage{
timeCeiling(timeValue, timeIntervalValue = 15,
  timeIntervalUnits = "minutes")
}
\arguments{
\item{timeValue}{typically a column of time Values to take the floor function}

\item{timeIntervalValue}{the value of the time interval}

\item{timeIntervalUnits}{the units of the time interval}
}
\description{
This function acts like a ceiling function for time.
}
\examples{

library(EmissionsHelper)
mytime <- reformatTime(c("2018-1-1 12:01", "2018-1-1 12:02",
                       "2018-1-1 12:16", "2018-1-1 12:17")) 
timeCeiling(mytime, 15, "minutes")                       


}
\keyword{time}
