% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeFloor.R
\name{timeFloor}
\alias{timeFloor}
\title{timeFloor}
\usage{
timeFloor(timeValue, timeIntervalValue = 15, timeIntervalUnits = "minutes")
}
\arguments{
\item{timeValue}{typically a column of time Values to take the floor function}

\item{timeIntervalValue}{the value of the time interval}

\item{timeIntervalUnits}{the units of the time interval}
}
\description{
This function acts like a floor function execpt for time.
}
\examples{

}
\keyword{time}
