% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatTime.R
\name{reformatTime}
\alias{reformatTime}
\title{reformatTime}
\usage{
reformatTime(dateToReformat, timeIfSeparateFromDate = NULL,
  dateFormat = "american", tz = "Etc/GMT+6", ...)
}
\arguments{
\item{dateToReformat}{the date and or time an event occured}

\item{timeIfSeparateFromDate}{the time of an event as a character array}

\item{tz}{time zone to use}

\item{format}{format string to output results}
}
\description{
This function is a wrapper for the as.POSIXct function. This function is 
designed require less effort.
}
\details{
The format parameter can be identical to the as.POSIXct function parameter
or it can be simplified to simply "american", "british", or "international".

The function identifies the format from knowledge of the order.

The user identifies which format order.  This code will issue a warning 
if an obvious issue occurs with the format.  If the dataset is large
getting the format is critical as testing is limited to the first 1000 datapoints. 
  
format = "american"      ... date order is (1) mm (2) day (3) year
format = "british"       ... date order is (1) day (2) mm (3) year
format = "international" ... date order is (1) year (2) mm (3) day

The time of day may not be part of the character string included with the
date.  If it is not then the timeIfSeparateFromDate paramter can include 
a text string with the time.

This function does not work if the data is already converted to a 
POSIXct time format.
}
\examples{
 
   reformatTime()
}
\keyword{time}
