% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameAllDataframeColumns.R
\name{renameAllDataframeColumns}
\alias{renameAllDataframeColumns}
\title{renameAllDataframeColumns}
\usage{
renameAllDataframeColumns(dataframe, newColumnNames)
}
\arguments{
\item{dataframe}{the dataframe to edit}

\item{newColumnNames}{the new column names}
}
\value{
same dataframe except with columns renamed
}
\description{
This function renames all the columns in a dataframe.
}
\details{
This function is useful because wind measuring equipment may output a less than
ideal format for R.  If column names are less than ideal for use in R 
this function can be used to change the names of all of the columns.

The first parameter is the dataframe

The second parameter is the column names to change the header.
}
\examples{

df <- data.frame(Column1 = 1, Column2 = 2)
renameAllDataframeColumns(df, c('ColumnA', 'ColumnB'))

}
\keyword{column}
\keyword{dataframe}
\keyword{rename}
