% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameDataframeColumn.R
\name{renameDataframeColumn}
\alias{renameDataframeColumn}
\title{renameDataframeColumn}
\usage{
renameDataframeColumn(dataframe, oldColumnName, newColumnName)
}
\arguments{
\item{dataframe}{the dataframe to edit}

\item{oldColumnName}{the old column name (Can also be the number of the column)}

\item{newColumnName}{the new name to change the old name to}
}
\value{
same dataframe except with columns renamed
}
\description{
This function renames a single column in R.
}
\details{
The first parameter is the dataframe name to change

The second column name is either the old column name or the index of the column.

The third is the name to change the dataframe to
}
\examples{

df <- data.frame(Column1 = 1, Column2 = 2)
df <- renameDataframeColumn(df, 'Column1', 'ColumnX')
df <- renameDateframeColumn(df, 2, 'ColumnY')


}
\keyword{column}
\keyword{dataframe}
\keyword{rename}
