% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeFloor.R
\name{timeFloor}
\alias{timeFloor}
\title{timeFloor}
\usage{
timeFloor(timeValue, timeIntervalValue = 15, timeIntervalUnits = "minutes")
}
\arguments{
\item{timeValue}{typically a column of time Values to take the floor function}

\item{timeIntervalValue}{the value of the time interval}

\item{timeIntervalUnits}{the units of the time interval}
}
\description{
This function acts like a floor function execpt for time.
}
\examples{

library(EmissionsHelper)
mytime <- reformatTime(c("2018-1-1 12:01", "2018-1-1 12:02",
                       "2018-1-1 12:16", "2018-1-1 12:17")) 
timeFloor(mytime, 15, "minutes") 

}
\keyword{time}
